+++ packages/react-frontend/src/SearchBar.css
