--- packages/react-frontend/src/components/Result.css
@@ -1,7 +1,0 @@
-.search-result {
-    padding: 10px 20px;
-}
-
-.search-result:hover {
-    background-color: #efefef;
-}
