--- packages/react-frontend/src/components/Result.jsx
@@ -1,11 +1,0 @@
-import React from 'react'
-
-export const Result = ((result) => {
-    return (
-        <div className="search-result" onClick={(e) => alert(`You selected ${result}!`)}>
-        
-        {result.title}
-   
-        </div>
-    );
-})
