--- packages/react-frontend/src/components/SearchResults.css
@@ -1,11 +1,0 @@
-.results-list {
-    width: 100%;
-    background: white;
-    display: flex;
-    flex-direction: column;
-    box-shadow: 0px 0px 8px #ddd;
-    border-radius: 10px;
-    margin-top: 1rem;
-    max-height: 300px;
-    overflow: scroll;
-}
